Refactor file paths for layout and public binds cards components
