+++ next/app/ui/layout/navigation/navigation-shimmer.tsx
