+++ next/app/ui/layout/public-binds-cards/cards-shimmer/index.tsx
@@ -1,4 +1,4 @@
-import SingleCardShimmer from "@/components/shared/cards/single-card-shimmer";
+import SingleCardShimmer from "@/app/ui/shared/cards/single-card-shimmer";
 
 export default function CardsShimmer() {
   return Array.from({ length: 9 }).map((_, index) => (
