+++ next/app/ui/layout/public-binds-cards/cards-shimmer/styles.module.css
