+++ next/app/ui/shared/auto-theme-switch.tsx
